refactor(bin): hide cmd package
it's only for internal use, ever
